relocated source to ./eligbility_test and eliminated sys.path inserts
